+++ protected/humhub/docs/CHANGELOG.md
@@ -13,6 +13,7 @@ HumHub Change Log
 - Fix #2571: last_login not set after registration direct login 
 - Enh: Always trigger dom widget events for widget `fire` until `triggerDom` is set to false
 - Enh: Added `richtextPaste` event
+- Enh: On search index rebuilding - use batch queries 
 
 1.2.0 (April 16, 2017)
 --------------------------------